# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Kasragay API
  description: |-
    ## Notes 
    ---
    - **09202400120** phone number is signed up by default as admin, and client.
      - **Admin ID** 89950e97-6b0f-4ef4-977a-8378b14bc4a7
      - **Client ID** 779033a2-4eaa-4817-aa81-24e24bd419f5
    - **09202400120-9** phone numbers are by-passed and available for testing.

    ### Logs
    ---
    **2025-07-09**
    - Initial commit
  version: 1.0.0
servers:
  - url: https://api.kasragay.com/v1
    description: API Server
...
paths:
...
  /user:
    get:
      tags:
        - user
      summary: Get user profile (10 r/m)
      description: Get user profile
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: query
          required: true
          description: User ID
          schema:
            type: string
            example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
            format: uuid
        - name: user_type
          in: query
          required: true
          description: User Type of user
          schema:
            type: string
            example: "client"
            enum: [client, admin]
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtUnauthorizedResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameNotFoundResponse"
        "410":
          description: User has been deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletedResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    put:
      tags:
        - user
      summary: Update user profile (10 r/m)
      description: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUserPutRequest"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserUserPutResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BadRequestResponse"
                  - $ref: "#/components/schemas/BadAvatarResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtUnauthorizedResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    delete:
      tags:
        - user
      summary: Delete user profile (10 r/m)
      description: Delete user profile (token is optional because of admin previlages)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUserDeleteRequest"
      responses:
        "200":
          description: Successful operation
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtUnauthorizedResponse"
        "403":
          description: Invalid Host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidHostResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundResponse"
        "410":
          description: User has been deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDeletedResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserUserPutRequest:
      type: object
      required:
        - id
        - username
        - name
        - avatar
        - user_type
      properties:
        id:
          type: string
          example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
          format: uuid
        username:
          type: string
          example: "Kasra"
          pattern: ^(?!.*\.\.)[a-zA-Z0-9\-_]+(\.[a-zA-Z0-9\-_]+)*$
          minLength: 3
          maxLength: 64
        name:
          type: string
          example: "John"
          pattern: ^[\x{0600}-\x{06FF}\x{FB50}-\x{FDFF}a-zA-Z\s-]{2,250}$
        avatar:
          type: string
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAPoAAAD6CAIAAAAHjs1qAAAB70lEQVR42uzSAQkAAAjEQBH7V9Ye/l2EsdmCFC0Bdge7g93B7mB3sDvYHewOdge7g92xO9gd7A52B7uD3cHuYHewO9gd7I7dwe5gd7A72B3sDnYHu4Pdwe5gd+wOdge7g93B7mB3sDvYHewOdge7Y3ewO9gd7A52B7uD3cHuYHewO9gd7I7dwe5gd7A72B3sDnYHu4Pdwe5gd+wOdge7g93B7mB3sDvYHewOdge7Y3ewO9gd7A52B7uD3cHuYHewO9gdu4Pdwe5gd7A72B3sDnYHu4Pdwe7YHewOdge7g93B7mB3sDvYHewOdge7Y3ewO9gd7A52B7uD3cHuYHewO9gdu4Pdwe5gd7A72B3sDnYHu4Pdwe7YHewOdge7g93B7mB3sDvYHewOdsfuYHewO9gd7A52B7uD3cHuYHewO9gdu4Pdwe5gd7A72B3sDnYHu4Pdwe7YHewOdge7g93B7mB3sDvYHewOdsfuYHewO9gd7A52B7uD3cHuYHewO3YHu4Pdwe5gd7A72B3sDnYHu4PdsTvYHewOdge7g93B7mB3sDvYHewOdsfuYHewO9gd7A52B7uD3cHuYHewO3YHu4Pdwe5gd7A72B3sDnYHu4PdsTvYHewOdge7g93B7mB3sDvYHeyO3cHu8MsFAAD//4g9AvbaMQNEAAAAAElFTkSuQmCC"
          format: base64
          description: Base64-encoded JPG or JPG 250x250px image
        user_type:
          type: string
          example: "client"
          enum: [admin, client]
    UserUserPutResponse:
      type: object
      properties:
        avatar:
          type: string
          example: "https://api.kasragay.com/v1/s3/avatars/client/f4f6a6b8-d625-44f7-befd-e10fd47705fd.png"
          description: Url of avatar's image
    UserUserDeleteRequest:
      type: object
      required:
        - id
        - user_type
      properties:
        id:
          type: string
          example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
          format: uuid
        user_type:
          type: string
          example: "client"
          enum: [admin, client]
        token:
          type: string
          example: "12345"
          pattern: ^\d{5}$
    User:
      type: object
      required:
        - id
        - username
        - name
        - user_type
      properties:
        id:
          type: string
          example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
          format: uuid
        username:
          type: string
          example: "Kasra"
          pattern: ^(?!.*\.\.)[a-zA-Z0-9\-_]+(\.[a-zA-Z0-9\-_]+)*$
          minLength: 3
          maxLength: 64
        name:
          type: string
          example: "John"
          pattern: ^[\x{0600}-\x{06FF}\x{FB50}-\x{FDFF}a-zA-Z\s-]{2,250}$
        avatar:
          type: string
          example: "https://api.kasragay.com/v1/s3/avatars/client/f4f6a6b8-d625-44f7-befd-e10fd47705fd.png"
          description: Url of avatar's image
        user_type:
          type: string
          example: "client"
          enum: [admin, client]
...
    TooManyRequestsResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 999
          enum: [999]
        message:
          type: string
          example: "too many requests"
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
    InternalServerErrorResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1000
          enum: [1000]
        message:
          type: string
          example: "internal server error"
...
    InvalidHostResponse:
      type: object
      required:
        - status
        - code
        - message
        - reasons
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1002
          enum: [1002]
        message:
          type: string
          example: "inavlid host"
        reasons:
          type: object
          required:
            - host
          properties:
            host:
              type: string
              example: "88.101.15.11:443"
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    BadRequestResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1004
          enum: [1004]
        message:
          type: string
          example: "bad request"
        reasons:
          type: object
          properties:
            field1:
              type: string
              example: "value1"
            field2:
              type: string
              example: "value2"

        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    UserNotFoundResponse:
      type: object
      required:
        - status
        - code
        - message
        - reasons
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1006
          enum: [1006]
        message:
          type: string
          example: "user not found"
        reasons:
          type: object
          properties:
            id:
              type: string
              example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
              format: uuid
            phone_number:
              type: string
              example: "09202400120"
              pattern: ^0\d{10}$
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
    UsernameNotFoundResponse:
      type: object
      required:
        - status
        - code
        - message
        - reasons
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1007
          enum: [1007]
        message:
          type: string
          example: "username not found"
        reasons:
          type: object
          properties:
            id:
              type: string
              example: "f4f6a6b8-d625-44f7-befd-e10fd47705fd"
              format: uuid
            phone_number:
              type: string
              example: "09202400120"
              pattern: ^0\d{10}$
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    JwtUnauthorizedResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1010
          enum: [1010]
        message:
          type: string
          example: "jwt unauthorized"
        reasons:
          type: object
          properties:
            type:
              type: string
              example: access
              enum: [access, refresh]
            wanted_type:
              type: string
              example: refresh
              enum: [access, refresh]
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    BadAvatarResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1013
          enum: [1013]
        message:
          type: string
          example: "bad avatar image"
        reasons:
          type: object
          properties:
            width:
              type: integer
              example: 5001
            height:
              type: integer
              example: 5001
            size:
              type: integer
              example: 10485761
            max_width:
              type: integer
              example: 5000
            max_height:
              type: integer
              example: 5000
            max_size:
              type: integer
              example: 10485760
            format:
              type: string
              example: "ai"
            valid_formats:
              type: array
              example: "['jpg', 'jpeg', 'png']"
            error:
              type: string
              example: "error decoding image"
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...
    UserDeletedResponse:
      type: object
      required:
        - status
        - code
        - message
      properties:
        status:
          type: string
          example: "error"
          enum: [error]
        code:
          type: integer
          example: 1015
          enum: [1015]
        message:
          type: string
          example: "user has been deleted"
        callers:
          type: array
          example: ["stack.svc1.method1", stack.svc2.method1, ...]
...